Refactor: Modularize SimpleJSPlugin for better maintainability
