--- REFACTORING_ANALYSIS.md
@@ -1,223 +1,0 @@
-# SimpleJS Plugin Refactoring Analysis
-
-## 📊 Before vs After Comparison
-
-| **Metric** | **Before (Monolithic)** | **After (Modular)** | **Improvement** |
-|------------|--------------------------|----------------------|-----------------|
-| **Main class lines** | 425 lines | ~120 lines | **71% reduction** |
-| **Single Responsibility** | ❌ Multiple concerns | ✅ Orchestration only | **Clear separation** |
-| **Testability** | ❌ Hard to unit test | ✅ Each component testable | **Fully testable** |
-| **Maintainability** | ❌ Complex changes | ✅ Isolated changes | **Easy maintenance** |
-| **Reusability** | ❌ Monolithic | ✅ Components reusable | **High reusability** |
-| **Code organization** | ❌ One giant file | ✅ Logical modules | **Well organized** |
-
-## 🏗️ New Architecture
-
-### **1. DataReader** (87 lines)
-**Responsibility:** File reading and parsing
-- ✅ YAML parsing with custom structure support
-- ✅ TSV parsing with type coercion  
-- ✅ Path resolution (relative/absolute)
-- ✅ Error handling with context
-
-```typescript
-const reader = new DataReader(app);
-const data = await reader.readYaml("data.yml", "current/file.md");
-```
-
-### **2. PlotlyManager** (92 lines)  
-**Responsibility:** Plotly integration and theming
-- ✅ CDN loading on-demand
-- ✅ Theme integration (light/dark/auto)
-- ✅ Cache management
-- ✅ Error recovery
-
-```typescript
-const plotly = new PlotlyManager(settings);
-await plotly.createPlot(container, { data, layout });
-```
-
-### **3. UIRenderer** (159 lines)
-**Responsibility:** HTML generation and styling  
-- ✅ Output containers with professional styling
-- ✅ Table rendering with hover effects
-- ✅ Console output formatting
-- ✅ Error/success messages
-- ✅ Control bars and buttons
-
-```typescript
-const ui = new UIRenderer();
-const container = ui.createOutputContainer(element);
-ui.createTable(container.content, rows, options);
-```
-
-### **4. CodeExecutor** (104 lines)
-**Responsibility:** Sandboxed code execution
-- ✅ Secure sandbox environment
-- ✅ Timeout protection  
-- ✅ Code validation (syntax, security)
-- ✅ Error handling and recovery
-
-```typescript
-const executor = new CodeExecutor(settings);
-await executor.execute(code, context);
-```
-
-### **5. ExecutionContextProvider** (78 lines)
-**Responsibility:** Runtime API creation
-- ✅ Creates the `readYaml`, `readTsv`, `plot`, `table`, `console` functions
-- ✅ Orchestrates component interactions
-- ✅ Console proxy with formatting
-
-```typescript
-const provider = new ExecutionContextProvider(reader, plotly, ui, settings);
-const context = provider.createContext(container, "current/file.md");
-```
-
-### **6. SimpleJSPlugin** (120 lines) 
-**Responsibility:** Plugin lifecycle and orchestration
-- ✅ Component initialization
-- ✅ Code block processor registration
-- ✅ Settings and commands
-- ✅ Clean lifecycle management
-
-## 🎯 Key Improvements
-
-### **1. Single Responsibility Principle**
-```typescript
-// Before: One class doing everything
-class SimpleJSPlugin {
-    parseYaml() { /* 50 lines */ }
-    parseTsv() { /* 40 lines */ }
-    createPlot() { /* 80 lines */ }
-    renderTable() { /* 60 lines */ }
-    executeCode() { /* 30 lines */ }
-    // ... 425 total lines
-}
-
-// After: Each component has one job
-class DataReader { /* Only file I/O */ }
-class PlotlyManager { /* Only plotting */ }  
-class UIRenderer { /* Only UI */ }
-```
-
-### **2. Improved Testability**
-```typescript
-// Before: Hard to test individual features
-test('plugin works', () => {
-    // Must test entire plugin at once
-    const plugin = new SimpleJSPlugin();
-    // Complex setup required...
-});
-
-// After: Easy to test each component
-test('YAML parsing works', () => {
-    const reader = new DataReader(mockApp);
-    const result = reader.parseYaml('key: value');
-    expect(result).toEqual({ key: 'value' });
-});
-
-test('table rendering works', () => {
-    const ui = new UIRenderer();
-    ui.createTable(mockContainer, [{ name: 'test' }]);
-    // Assert DOM structure...
-});
-```
-
-### **3. Configuration and Extensibility**
-```typescript
-// Easy to add new data formats
-class DataReader {
-    async readCsv(path: string) { /* new format */ }
-    async readJson(path: string) { /* another format */ }
-}
-
-// Easy to add new visualization types
-class PlotlyManager {
-    async create3DPlot(spec: Plot3DSpec) { /* new plot type */ }
-    async createAnimation(spec: AnimationSpec) { /* animated plots */ }
-}
-```
-
-### **4. Better Error Handling**
-```typescript
-// Each component handles its own errors appropriately
-try {
-    await dataReader.readYaml(path);
-} catch (error) {
-    throw new YamlParseError(`Invalid YAML in ${path}: ${error.message}`);
-}
-
-try {
-    await plotlyManager.createPlot(spec);
-} catch (error) {
-    throw new PlotRenderError(`Failed to render plot: ${error.message}`);
-}
-```
-
-### **5. Memory Management**
-```typescript
-// Clean resource cleanup
-onunload() {
-    this.plotlyManager?.clearCache();  // Clear Plotly resources
-    // Other components clean themselves up
-}
-```
-
-## 🚀 Development Benefits
-
-### **Parallel Development**
-- Different developers can work on different components
-- No merge conflicts between unrelated features
-- Easier code reviews (smaller, focused changes)
-
-### **Feature Addition**  
-```typescript
-// Adding new data source? Only touch DataReader
-class DataReader {
-    async readParquet(path: string) { /* new method */ }
-}
-
-// Adding new chart type? Only touch PlotlyManager  
-class PlotlyManager {
-    async createHeatmap(spec: HeatmapSpec) { /* new method */ }
-}
-```
-
-### **Bug Fixes**
-- Issues isolated to specific components
-- Easier to identify root cause
-- Changes don't affect unrelated functionality
-
-## 📦 File Structure
-
-```
-src/
-├── DataReader.ts           # File I/O and parsing
-├── PlotlyManager.ts        # Visualization engine  
-├── UIRenderer.ts           # HTML/CSS generation
-├── CodeExecutor.ts         # Sandboxed execution
-├── ExecutionContextProvider.ts  # Runtime API
-└── simple-settings.ts      # Settings interface
-
-main-refactored.ts          # Clean orchestration (120 lines)
-```
-
-## ✅ Migration Path
-
-1. **Install new components** (all files in `src/`)
-2. **Replace main.ts** with `main-refactored.ts`  
-3. **Test functionality** (same external API)
-4. **Extend as needed** (add features to appropriate components)
-
-## 🎯 Result
-
-**71% reduction in main class complexity** while gaining:
-- ✅ Better testability
-- ✅ Easier maintenance  
-- ✅ Component reusability
-- ✅ Cleaner architecture
-- ✅ Parallel development capability
-- ✅ Future extensibility
-
-The refactored version maintains **100% compatibility** with existing SimpleJS blocks while providing a much cleaner, more maintainable codebase.
